Refactor dynamic type extension for ODM use
* Use generic ObjectManager naming
* Make type extensions that differ for ODM more flexible
