+++ src/Type/Doctrine/EntityManagerFindDynamicReturnTypeExtension.php
@@ -2,53 +2,22 @@
 
 namespace PHPStan\Type\Doctrine;
 
-use PhpParser\Node\Expr\MethodCall;
-use PHPStan\Analyser\Scope;
-use PHPStan\Reflection\MethodReflection;
-use PHPStan\Type\Constant\ConstantStringType;
-use PHPStan\Type\MixedType;
-use PHPStan\Type\ObjectType;
-use PHPStan\Type\Type;
-use PHPStan\Type\TypeCombinator;
+use const E_USER_DEPRECATED;
+use function trigger_error;
 
-class EntityManagerFindDynamicReturnTypeExtension implements \PHPStan\Type\DynamicMethodReturnTypeExtension
+class EntityManagerFindDynamicReturnTypeExtension extends ObjectManagerFindDynamicReturnTypeExtension
 {
 
-	public function getClass(): string
+	public function __construct()
 	{
-		return 'Doctrine\Common\Persistence\ObjectManager';
-	}
-
-	public function isMethodSupported(MethodReflection $methodReflection): bool
-	{
-		return in_array($methodReflection->getName(), [
-			'find',
-			'getReference',
-			'getPartialReference',
-		], true);
-	}
-
-	public function getTypeFromMethodCall(
-		MethodReflection $methodReflection,
-		MethodCall $methodCall,
-		Scope $scope
-	): Type
-	{
-		$mixedType = new MixedType();
-		if (count($methodCall->args) === 0) {
-			return $mixedType;
-		}
-		$argType = $scope->getType($methodCall->args[0]->value);
-		if (!$argType instanceof ConstantStringType) {
-			return $mixedType;
-		}
-
-		$type = new ObjectType($argType->getValue());
-		if ($methodReflection->getName() === 'find') {
-			$type = TypeCombinator::addNull($type);
-		}
-
-		return $type;
+		@trigger_error(
+			sprintf(
+				'Class %s is deprecated and will be removed. Use %s instead',
+				self::class,
+				ObjectManagerFindDynamicReturnTypeExtension::class
+			),
+			E_USER_DEPRECATED
+		);
 	}
 
 }
