+++ src/Type/Doctrine/EntityManagerGetRepositoryDynamicReturnTypeExtension.php
@@ -2,52 +2,24 @@
 
 namespace PHPStan\Type\Doctrine;
 
-use PhpParser\Node\Expr\MethodCall;
-use PHPStan\Analyser\Scope;
-use PHPStan\Reflection\MethodReflection;
-use PHPStan\Reflection\ParametersAcceptorSelector;
-use PHPStan\Type\Constant\ConstantStringType;
-use PHPStan\Type\MixedType;
-use PHPStan\Type\Type;
+use const E_USER_DEPRECATED;
+use function trigger_error;
 
-class EntityManagerGetRepositoryDynamicReturnTypeExtension implements \PHPStan\Type\DynamicMethodReturnTypeExtension
+class EntityManagerGetRepositoryDynamicReturnTypeExtension extends ObjectManagerGetRepositoryDynamicReturnTypeExtension
 {
-
-	/** @var string */
-	private $repositoryClass;
 
 	public function __construct(string $repositoryClass)
 	{
-		$this->repositoryClass = $repositoryClass;
-	}
+		@trigger_error(
+			sprintf(
+				'Class %s is deprecated and will be removed. Use %s instead',
+				self::class,
+				ObjectManagerGetRepositoryDynamicReturnTypeExtension::class
+			),
+			E_USER_DEPRECATED
+		);
 
-	public function getClass(): string
-	{
-		return 'Doctrine\Common\Persistence\ObjectManager';
-	}
-
-	public function isMethodSupported(MethodReflection $methodReflection): bool
-	{
-		return $methodReflection->getName() === 'getRepository';
-	}
-
-	public function getTypeFromMethodCall(
-		MethodReflection $methodReflection,
-		MethodCall $methodCall,
-		Scope $scope
-	): Type
-	{
-		if (count($methodCall->args) === 0) {
-			return ParametersAcceptorSelector::selectSingle(
-				$methodReflection->getVariants()
-			)->getReturnType();
-		}
-		$argType = $scope->getType($methodCall->args[0]->value);
-		if (!$argType instanceof ConstantStringType) {
-			return new MixedType();
-		}
-
-		return new EntityRepositoryType($argType->getValue(), $this->repositoryClass);
+		parent::__construct('Doctrine\Common\Persistence\ObjectManager', $repositoryClass);
 	}
 
 }
