+++ src/Type/Doctrine/ObjectManagerFindDynamicReturnTypeExtension.php
@@ -1,0 +1,54 @@
+<?php declare(strict_types = 1);
+
+namespace PHPStan\Type\Doctrine;
+
+use PhpParser\Node\Expr\MethodCall;
+use PHPStan\Analyser\Scope;
+use PHPStan\Reflection\MethodReflection;
+use PHPStan\Type\Constant\ConstantStringType;
+use PHPStan\Type\MixedType;
+use PHPStan\Type\ObjectType;
+use PHPStan\Type\Type;
+use PHPStan\Type\TypeCombinator;
+
+class ObjectManagerFindDynamicReturnTypeExtension implements \PHPStan\Type\DynamicMethodReturnTypeExtension
+{
+
+	public function getClass(): string
+	{
+		return 'Doctrine\Common\Persistence\ObjectManager';
+	}
+
+	public function isMethodSupported(MethodReflection $methodReflection): bool
+	{
+		return in_array($methodReflection->getName(), [
+			'find',
+			'getReference',
+			'getPartialReference',
+		], true);
+	}
+
+	public function getTypeFromMethodCall(
+		MethodReflection $methodReflection,
+		MethodCall $methodCall,
+		Scope $scope
+	): Type
+	{
+		$mixedType = new MixedType();
+		if (count($methodCall->args) === 0) {
+			return $mixedType;
+		}
+		$argType = $scope->getType($methodCall->args[0]->value);
+		if (!$argType instanceof ConstantStringType) {
+			return $mixedType;
+		}
+
+		$type = new ObjectType($argType->getValue());
+		if ($methodReflection->getName() === 'find') {
+			$type = TypeCombinator::addNull($type);
+		}
+
+		return $type;
+	}
+
+}
